Split data before doing in-depth analysis
This prevents forms of leakage in visual inspection and features based on global statistics. Splitting here is not particularly critical, results are mostly the same, and this version of the data would not be used to train the model anyway. By splitting here, however, we are completely consistent with the data that the model will actually see.