Add a simple caching system for unique query ids
aiocache provides an asyncio-safe simple set of features to handle a
memory cache. It is a fronted to various well known backends, which is
very handy. The current setup (no backend specified) is a simple wrapper
to an asincio locked dictionary, which would not be shared between
FastAPI workers. It is then suitable for this study case, where a single
worker is used. However, with minimal effort it is possible to swap e.g.
to the Redis backend, which would be a global instance shared between
workers. And this comes without changes in the code interface.