Add model prediction endpoints POST and DELETE
POST to create a query. The system will return a new endpoint,
which is the location of the query resource. Use DELETE on this very
location to remove the entire query (for the scope of this project,
there is no automatic expiration, users are supposed to delete their
queries when they are done).